file structure for signin/up hooks adjusted
