+++ client/src/hooks/Signin/useBuyerSignin.js
@@ -1,30 +1,30 @@
-import Auth from '../utils/auth';
+import Auth from '../../utils/auth';
 import { useState } from 'react';
 import { useMutation } from '@apollo/client';
 import { useNavigate } from 'react-router-dom';
-import { buyer_Signup } from '../utils/mutations';
-import { useAuthContext } from './useAuthContext';
+import { buyer_login } from '../../utils/mutations';
+import { useAuthContext } from '../useAuthContext';
 
-export const useBuyerSignup = () => {
+export const useBuyerSignin = () => {
     const [stateError, setStateError] = useState(null)
     const [isLoading, setIsLoading] = useState(null)
     const {dispatch} = useAuthContext()
     const navigate = useNavigate()
 
     //  Mutation
-    const [AddUser, { error, loading, data }] = useMutation(buyer_Signup);
+    const [LoginUser, { error, loading, data }] = useMutation(buyer_login);
 
-    const signup = async (formState) => {
+    const signin = async (formState) => {
         setIsLoading(true)
         setStateError(null)
 
         try {
-            const { data } = await AddUser({
+            const { data } = await LoginUser({
             variables: formState,
             });
             
-            // create new token
-            Auth.login(data.AddUser.token);
+            // login with users token
+            Auth.login(data.Userlogin.token);
             
             // update the auth context
             dispatch({ type: 'LOGIN', payload: data })
@@ -39,9 +39,9 @@ export const useBuyerSignup = () => {
       } catch (e) {
             setStateError(true)
             setIsLoading(false)
-            console.error('AddUser error in useSignup() hook:', e);
+            console.error('LoginUser error in useSignin() hook: ', e);
       }
          
     }
-    return { signup, isLoading, stateError }
+    return { signin, isLoading, stateError }
 }
