+++ client/src/pages/Signup/Buyer/BuyerSignup.jsx
@@ -11,38 +11,33 @@ import { Alert } from '@mui/material';
 import Typography from '@mui/material/Typography';
 import Container from '@mui/material/Container';
 import ShoppingBasketIcon from '@mui/icons-material/ShoppingBasket';
-import { useBuyerSignup } from '../../../hooks/useBuyerSignup';
+import { useBuyerSignup } from '../../../hooks/Signup/useBuyerSignup';
 import { useAuthContext } from '../../../hooks/useAuthContext';
 
-
 export default function BuyerSignup() {
-  const { user } = useAuthContext()
-  const { signup, stateError, isLoading } = useBuyerSignup() // custom hook
-  const navigate = useNavigate()
-
+  const { user } = useAuthContext();
+  const { signup, stateError, isLoading } = useBuyerSignup(); // custom hook
+  const navigate = useNavigate();
 
   // Error & Alert States
   const [showSuccessAlert, setShowSuccessAlert] = useState(false);
   const [showErrorAlert, setShowErrorAlert] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
 
-
-   // Form state
-   const [formState, setFormState] = useState({
+  // Form state
+  const [formState, setFormState] = useState({
     firstName: '',
     lastName: '',
     email: '',
     password: '',
   });
 
-  
   // On render, check if logged in => Send to profile page
   useEffect(() => {
     if (user) {
       navigate('/profile');
     }
   }, []);
-
 
   // OnChange - update form state
   const handleChange = (event) => {
@@ -53,11 +48,11 @@ export default function BuyerSignup() {
     });
   };
 
-    // onClose - clear error message 
-    const handleClearError = () => {
-      setErrorMessage('');
-      setShowErrorAlert(false);
-    };
+  // onClose - clear error message
+  const handleClearError = () => {
+    setErrorMessage('');
+    setShowErrorAlert(false);
+  };
 
   // OnSubmit - validation check + run signup() hook
   const handleSubmit = async (event) => {
@@ -77,7 +72,9 @@ export default function BuyerSignup() {
     }
 
     if (!emailRegex.test(formState.email) && formState.password.length < 8) {
-      setErrorMessage('The email address is invalid, and your password must be at least 8 characters long')
+      setErrorMessage(
+        'The email address is invalid, and your password must be at least 8 characters long'
+      );
     }
 
     if (errorMessage) {
@@ -87,14 +84,13 @@ export default function BuyerSignup() {
 
     try {
       // console.log('Signup Form state:', formState);
-      await signup(formState)
+      await signup(formState);
       setShowSuccessAlert(true);
     } catch (e) {
       setShowErrorAlert(true);
       console.error('signup() error in BuyerSignup:', e);
     }
   };
-
 
   return (
     <Container component='main' maxWidth='xs'>
